offline mode v1

+++ app/src/main/java/com/dezeta/guessit/ui/main/MainActivity.kt
@@ -1,21 +1,28 @@
 package com.dezeta.guessit.ui.main
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
 import android.content.res.Configuration
 import android.graphics.Color
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
+import android.view.animation.Animation
+import android.view.animation.AnimationUtils
 import android.widget.Button
 import android.widget.ImageView
+import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.DrawableCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.navigation.NavController
@@ -24,16 +31,13 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import androidx.work.OneTimeWorkRequestBuilder
 import com.airbnb.lottie.LottieAnimationView
 import com.bumptech.glide.Glide
 import com.dezeta.guessit.R
 import com.dezeta.guessit.databinding.ActivityMainBinding
-import com.dezeta.guessit.domain.entity.Guess
 import com.dezeta.guessit.ui.login.LoginActivity
 import com.dezeta.guessit.utils.CloudStorageManager
 import com.dezeta.guessit.utils.Locator
-import com.dezeta.guessit.utils.MyWorker
 import com.google.android.material.navigation.NavigationView
 import com.google.firebase.auth.FirebaseAuth
 import de.hdodenhof.circleimageview.CircleImageView
@@ -42,14 +46,13 @@ import kotlin.properties.Delegates
 
 class MainActivity : AppCompatActivity() {
     private lateinit var lottieAnimationView: LottieAnimationView
-
     private lateinit var manager: CloudStorageManager
     private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var navController: NavController
     private lateinit var binding: ActivityMainBinding
     private lateinit var navView: NavigationView
     private var isDarkThemeOn by Delegates.notNull<Boolean>()
-
+    lateinit var fadeOutAnimation: Animation
 
     private lateinit var preferences: SharedPreferences.Editor
     private val viewModel: ViewModelMain by viewModels()
@@ -67,14 +70,14 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         isDarkThemeOn =
             (resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK) == Configuration.UI_MODE_NIGHT_YES
-        setTemes()
 
+        setTemes()
+        fadeOutAnimation = AnimationUtils.loadAnimation(this, R.anim.fade_out)
         binding = ActivityMainBinding.inflate(layoutInflater)
 
         setContentView(binding.root)
         preferences =
             getSharedPreferences(getString(R.string.prefs_file), Context.MODE_PRIVATE).edit()
-
 
         setSupportActionBar(binding.appBarMain.toolbar)
 
@@ -96,44 +99,91 @@ class MainActivity : AppCompatActivity() {
     }
 
     fun updateHeader() {
-        viewModel.loadUser(Locator.email)
+        if (isOnline())
+            viewModel.loadUser(Locator.email)
+
+    }
+
+    fun setOflineHeader() {
+        val navigationView = findViewById<View>(R.id.nav_view) as NavigationView
+        val headerView: View = navigationView.getHeaderView(0)
+        val btnDeleteProfile = headerView.findViewById<Button>(R.id.navBtnDeleteProfile)
+        val imgEditProfile = headerView.findViewById<ImageView>(R.id.navImgEditProfile)
+        val tvDrawerName = headerView.findViewById<TextView>(R.id.navTvName)
+        val tvEmail = headerView.findViewById<TextView>(R.id.navTvEmail)
+        val navImg = headerView.findViewById<CircleImageView>(R.id.navImgProfile)
+        val navLinerLayout = headerView.findViewById<LinearLayout>(R.id.navLinearLayout)
+        val navImgDolar = headerView.findViewById<ImageView>(R.id.navImgDolar)
+
+        navImg.setImageResource(R.drawable.user_profile)
+        tvEmail.text = ContextCompat.getString(this, R.string.nav_email)
+        tvDrawerName.text = ContextCompat.getString(this, R.string.nav_name_user)
+
+        btnDeleteProfile.isEnabled = false
+
+        imgEditProfile.visibility = View.INVISIBLE
+        navImgDolar.visibility = View.INVISIBLE
+        navLinerLayout.visibility = View.INVISIBLE
+
+
     }
 
     private fun setupHeader() {
         val navigationView = findViewById<View>(R.id.nav_view) as NavigationView
         val headerView: View = navigationView.getHeaderView(0)
-
+        val btnDeleteProfile = headerView.findViewById<Button>(R.id.navBtnDeleteProfile)
+        val imgEditProfile = headerView.findViewById<ImageView>(R.id.navImgEditProfile)
         lottieAnimationView =
             headerView.findViewById<LottieAnimationView>(R.id.navLottieLoadAnimation)
-        val btnEditProfile = headerView.findViewById<Button>(R.id.navBtnDeleteProfile)
-        val imgEditProfile = headerView.findViewById<ImageView>(R.id.navImgEditProfile)
         val tvDrawerName = headerView.findViewById<TextView>(R.id.navTvName)
         val tvEmail = headerView.findViewById<TextView>(R.id.navTvEmail)
         val tvDrawerPoint = headerView.findViewById<TextView>(R.id.navTvPoint)
         val tvPrePoint = headerView.findViewById<TextView>(R.id.navTvPrePoint)
         val tvPostPoint = headerView.findViewById<TextView>(R.id.navTvPostPoint)
+        val navLinerLayout = headerView.findViewById<LinearLayout>(R.id.navLinearLayout)
+        val navImgDolar = headerView.findViewById<ImageView>(R.id.navImgDolar)
+
+
         if (isDarkThemeOn) {
             tvPrePoint.setTextColor(Color.WHITE)
             tvPostPoint.setTextColor(Color.WHITE)
         }
+        if (isOnline()) {
+            imgEditProfile.visibility = View.VISIBLE
+            navImgDolar.visibility = View.VISIBLE
+            navLinerLayout.visibility = View.VISIBLE
+            viewModel.getUserProfileImageByEmail(manager)
+            tvEmail.text = viewModel.user.value!!.email
+            tvDrawerName.text = viewModel.user.value!!.name
+            tvDrawerPoint.text = viewModel.user.value!!.point.toString()
+            imgEditProfile.setOnClickListener {
+                pickPhotoFromGallery()
+            }
 
-        viewModel.getUserProfileImageByEmail(manager)
-        tvEmail.text = viewModel.user.value!!.email
-        tvDrawerName.text = viewModel.user.value!!.name
-        tvDrawerPoint.text = viewModel.user.value!!.point.toString()
-        imgEditProfile.setOnClickListener {
-            pickPhotoFromGallery()
+            btnDeleteProfile.setOnClickListener {
+                showConfirmationDialog()
+            }
         }
 
-        btnEditProfile.setOnClickListener {
+    }
+
+    private fun showConfirmationDialog() {
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("¿Desea eliminar la cuenta?")
+        builder.setPositiveButton("Si") { _, _ ->
             with(lottieAnimationView) {
                 visibility = View.VISIBLE
                 setAnimation(R.raw.load_image)
                 playAnimation()
             }
             viewModel.deleteUser()
-
         }
+
+        builder.setNegativeButton("No") { _, _ ->
+        }
+        builder.show()
+
+
     }
 
     private fun setup() {
@@ -200,8 +250,8 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.menu_main, menu)
-        if(isDarkThemeOn){
-            tintMenuItemIcon(menu, R.id.action_signOut,Color.WHITE)
+        if (isDarkThemeOn) {
+            tintMenuItemIcon(menu, R.id.action_signOut, Color.WHITE)
         }
         return true
     }
@@ -220,6 +270,7 @@ class MainActivity : AppCompatActivity() {
             else -> super.onOptionsItemSelected(item)
         }
     }
+
     private fun tintMenuItemIcon(menu: Menu, idItem: Int, color: Int) {
         var drawable = menu.findItem(idItem).icon
         drawable = DrawableCompat.wrap(drawable!!)
@@ -251,6 +302,23 @@ class MainActivity : AppCompatActivity() {
 
 
     }
+
+    fun isOnline(): Boolean {
+        val connectivityManager =
+            getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val capabilities =
+            connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
+        if (capabilities != null) {
+            if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) ||
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) ||
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)
+            ) {
+                return true
+            }
+        }
+        return false
+    }
+
 }
